backend/requirements.txt
Django==4.2.3
Pillow==9.5.0
requests==2.28.2
django-resized==1.0.2
django-debug-toolbar==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.0
